Add provision output for storage product
