+++ public/extensions/demo/components/forms/fields.config.js
@@ -38,83 +38,6 @@
             }
         });
 
-        Forms.fields('demo_rds_admin_username', {
-            type: 'text',
-            templateOptions: {
-                label: 'Admin Username'
-            },
-            validators: {
-                firstCharacterIsLetter: {
-                    expression: function($viewValue, $modelValue, scope) {
-                        var value = $modelValue || $viewValue;
-                        return /^[A-Za-z].*/.test(value);
-                    },
-                    message: '"Username must start with a letter"'
-                },
-                isLongEnough: {
-                    expression: function($viewValue, $modelValue, scope) {
-                        var value = $modelValue || $viewValue;
-                        return /.{5,}/.test(value);
-                    },
-                    message: '"Username must be at least 5 characters long"'
-                },
-                noSpecialChars: {
-                    expression: function($viewValue, $modelValue, scope) {
-                        var value = $modelValue || $viewValue;
-                        return /^[A-Za-z0-9_]+$/.test(value);
-                    },
-                    message: '"Username can not contain any special characters"'
-                }
-            }
-        });
-
-        Forms.fields('demo_rds_admin_password', {
-            type: 'password',
-            templateOptions: {
-                label: 'Admin Password'
-            },
-            validators: {
-                isLongEnough: {
-                    expression: function($viewValue, $modelValue, scope) {
-                        var value = $modelValue || $viewValue;
-                        return /.{8,}/.test(value);
-                    },
-                    message: '"Password must be at least 8 characters long"'
-                },
-                noSpecialChars: {
-                    expression: function($viewValue, $modelValue, scope) {
-                        var value = $modelValue || $viewValue;
-                        return !/[\"\/\@]/.test(value);
-                    },
-                    message: '"Password can not contain \\"/\\", \\"\\"\\" or \\"@\\" "'
-                }
-            }
-        });
-
-        Forms.fields('demo_server_flavors', {
-            type: 'async_select',
-            templateOptions: {
-                label: 'Instance Type',
-                options: []
-            },
-            data: {
-                action: 'serverFlavors'
-            },
-            controller: DemoDataController
-        });
-
-        Forms.fields('demo_vpcs', {
-            type: 'async_select',
-            templateOptions: {
-                label: 'VPC',
-                options: []
-            },
-            data: {
-                action: 'vpcs'
-            },
-            controller: DemoDataController
-        });
-
         Forms.fields('demo_subnets', {
             type: 'select',
             templateOptions: {
@@ -151,42 +74,6 @@
             }
         });
 
-        Forms.fields('demo_zones', {
-            type: 'async_select',
-            templateOptions: {
-                label: 'Availability Zone',
-                options: []
-            },
-            data: {
-                action: 'zones'
-            },
-            controller: DemoDataController
-        });
-
-        Forms.fields('demo_key_names', {
-            type: 'async_select',
-            templateOptions: {
-                label: 'Key Name',
-                options: []
-            },
-            data: {
-                action: 'keyNames'
-            },
-            controller: DemoDataController
-        });
-
-        Forms.fields('demo_security_groups', {
-            type: 'async_select',
-            templateOptions: {
-                label: 'Security Group',
-                options: []
-            },
-            data: {
-                action: 'securityGroups'
-            },
-            controller: DemoDataController
-        });
-
         /** @ngInject */
         function DemoDataController($scope, DemoData, Toasts) {
             var provider = $scope.formState.provider;
@@ -212,15 +99,7 @@
                     });
                     $scope.to.loading = DemoData[action](provider.id, 'none').then(handleResults, handleError);
                     break;
-                case 'rdsVersions':
-                    $scope.$parent.$watch('model.engine' , function (newValue, oldValue, theScope) {
-                        if(newValue !== oldValue) {
-                            $scope.to.loading = DemoData[action](provider.id, newValue).then(handleResults, handleError);
-                        }
-                    });
-                    $scope.to.loading = DemoData[action](provider.id, 'none').then(handleResults, handleError);
-                    break;
-                case 'rdsFlavors':
+                case 'demoFlavors':
                     $scope.$parent.$watch('model.version' , function (newValue, oldValue, theScope) {
                         if(newValue !== oldValue) {
                             $scope.to.loading = DemoData[action](provider.id, $scope.model.engine, newValue).then(handleResults, handleError);
